Chemical Bonding

Chemical bonding interactions are represented in the bonds attribute of a Crystal as a graph where the nodes of the graph correspond to the crystal's atoms.

Xtals.BondingRule — Type

bonding_rule = BondingRule(:Ca, :O, 0.4, 2.0)
bonding_rules = [BondingRule(:H, :*, 0.4, 1.2),
                 BondingRule(:*, :*, 0.4, 1.9)]

A rule for determining if two atoms within a crystal are bonded.

Attributes

-species_i::Symbol: One of the atoms types for this bond rule -species_j::Symbol: The other atom type for this bond rule -min_dist: The minimum distance between the atoms for bonding to occur -max_dist: The maximum distance between the atoms for bonding to occur

Missing docstring.

Missing docstring for get_bonding_rules, set_bonding_rules. Check Documenter's build log for details.

Xtals.add_bonding_rules — Function

add_bonding_rules(bonding_rules)

Adds bonding_rules to the beginning of the global bonding rules list

Arguments

bonding_rules::Array{BondingRule} : the array of bonding rules to add

Xtals.read_bonding_rules — Function

read_bonding_rules("file.csv")

Reads a CSV file of bonding rules and returns a BondingRule array.

Arguments

filename::String : name of file in data directory containing bonding rules

Xtals.write_bonding_rules — Function

write_bonding_rules("file.csv")

Writes bonding rules to a CSV file that can be loaded with read_bonding_rules

Arguments

filename::String : The name of the output file

bonding_rules::Array{BondingRule} : (Optional) The rules to write to file. If not specified, the global rules are written.

Xtals.infer_bonds! — Function

infer_bonds!(crystal, include_bonds_across_periodic_boundaries,
                bonding_rules=[BondingRule(:H, :*, 0.4, 1.2), BondingRule(:*, :*, 0.4, 1.9)])

Populate the bonds in the crystal object based on the bonding rules. If a pair doesn't have a suitable rule then they will not be considered bonded.

:* is considered a wildcard and can be substituted for any species. It is a good idea to include a bonding rule between two :* to allow any atoms to bond as long as they are close enough.

The bonding rules are hierarchical, i.e. the first bonding rule takes precedence over the latter ones.

Arguments

-crystal::Crystal: The crystal that bonds will be added to -include_bonds_across_periodic_boundaries::Bool: Whether to check across the periodic boundary when calculating bonds -bonding_rules::Array{BondingRule, 1}: The array of bonding rules that will be used to fill the bonding information. They are applied in the order that they appear. -covalent_radii::Dict{Symbol, Dict{Symbol, Float64}}: Cordero parameters to use for calculating bonding rules. See covalent_radii -σ::Float64: Number of Cordero estimated standard deviations to use if calculating bonding rules from covalent radii. -min_tol::Float64: Minimum covalent radius tolerance if calculating bonding rules from covalent radii.

Xtals.bond_sanity_check — Function

sane_bonds = bond_sanity_check(crystal)

Run sanity checks on crystal.bonds.

is the bond graph fully connected? i.e. does every vertex (=atom) in the bond graph have at least one edge?

each hydrogen can have only one bond

each carbon can have a maximum of four bonds

if sanity checks fail, refer to write_bond_information to write a .vtk to visualize the bonds.

Print warnings when sanity checks fail. Return true if sanity checks pass, false otherwise.

Xtals.remove_bonds! — Function

remove_bonds!(crystal)

Remove all bonds from a crystal structure, crystal::Crystal.

Xtals.infer_geometry_based_bonds! — Function

infer_geometry_based_bonds!(crystal, include_bonds_across_periodic_boundaries::Bool)

Infers bonds by first finding which atoms share a Voronoi face, and then bond the atoms if the distance between them is less than the sum of the covalent radius of the two atoms (plus a tolerance).

Arguments

-crystal::Crystal: The crystal structure -include_bonds_across_periodic_boundaries::Bool: Whether to check across the periodic boundaries -r::Float: voronoi radius, Å -σ::Float: number of estimated standard deviations to use for covalent radius tolerance -min_tol::Float: minimum tolerance for calculated bond distances, Å -covalent_radii::Dict{Symbol, Dict{Symbol, Float64}}: See covalent_radii

Xtals.write_bond_information — Function

write_bond_information(crystal, filename)
write_bond_information(crystal, center_at_origin=false)

Writes the bond information from a crystal to the selected filename.

Arguments

-crystal::Crystal: The crystal to have its bonds written to a vtk file -filename::String: The filename the bond information will be saved to. If left out, will default to crystal name.

center_at_origin::Bool: center the coordinates at the origin of the crystal