Visualizing structures

Crystals can be visualized using Bio3DView and either Blink, Jupyter, or Pluto.

To include a visualization within a Pluto or Jupyter notebook, simply call view_crystal. To launch a visualization window directly from a script or the REPL, use Blink, like so:

using Blink
xtal = Crystal("SBMOF-1.cif")
infer_bonds!(xtal, true)
view_crystal(xtal)

Docs

Xtals.view_crystal — Function

`view_crystal(xtal, drop_cross_pb_bonds=true)`

Launch a GUI window displaying the crystal.

Arguments

xtal::Crystal : the crystal to display

drop_cross_pb_bonds::Bool : Optional. Set to true to remove bonds that extend across periodic boundaries of the unit cell (these tend to mess up the visualization). Defaults to true.